+++ talentmap_api/fsbid/services/employee.py
@@ -43,22 +43,6 @@ def get_employee_information(jwt_token, emp_id):
     except:
         return {}
 
-def get_user_information(jwt_token, perdet_seq_num):
-    '''
-    Gets the office_phone and office_address for the employee
-    '''
-    url = f"{SECREF_ROOT}/user?request_params.perdet_seq_num={perdet_seq_num}"
-    user = requests.get(url, headers={'JWTAuthorization': jwt_token, 'Content-Type': 'application/json'},
-                            verify=False).json()  # nosec
-    user = next(iter(user.get('Data', [])), {})
-    try:
-        return {
-            "office_address": user['gal_address_text'],
-            "office_phone": user['gal_phone_nbr_text'],
-        }
-    except:
-        return {}
-
 
 def map_group_to_fsbid_role(jwt_token):
     '''
